Add logic to handle mirror cards
Trello has added functionality for mirror cards [[1]], which allow cards
to be on two boards at once. The way this appears to work from the point
of view of the REST API is that a card can have the `role` `mirror`, and
then the `name` of the card points to the original.

Unfortunately it seems that adding an attachment to a mirror card via the
REST API changes the role of the card from `mirror`, breaking the link,
and making the card useless.

This commit fixes the behaviour of the Trello poster so that it checks
whether a card is a mirror card, and updates the card being mirrored if
necessary.

[1]: https://support.atlassian.com/trello/docs/mirroring-cards/