Update workflows to test the code in a PR
Currently we have a GitHub Actions workflow that runs the action in the
main branch of this repository for all PRs.

This makes it hard to test change to the action in a PR; either the code
has to be merged first, or the workflow configuration has to be
temporarily changed.

This PR change the workflow to always use the code in the branch for the
PR, so that GitHub Actions will test the changes in the PR more easily.

This is all subject to the security settings for GitHub Actions in this
repo, of course.